refactor: reorganize run and io modules
