+++ crates/floe-core/src/io/fs/mod.rs
@@ -1,0 +1,2 @@
+pub mod local;
+pub mod s3;
