+++ crates/floe-core/src/io/mod.rs
@@ -1,4 +1,3 @@
-pub mod read_csv;
-pub mod resolve;
-pub mod s3;
+pub mod fs;
+pub mod read;
 pub mod write;
