+++ crates/floe-core/src/io/read/mod.rs
@@ -1,0 +1,3 @@
+pub mod csv;
+
+pub use csv::{read_csv_file, read_csv_header, CsvReadPlan};
